requirements.txt
agate==1.7.1
alembic==1.16.2
anyio==4.9.0
attrs==25.3.0
babel==2.17.0
backoff==2.2.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
coloredlogs==14.0
croniter==6.0.0
dagster==1.6.6
dagster-dbt==0.22.6
dagster-graphql==1.6.6
dagster-pipes==1.6.6
dagster-webserver==1.6.6
dbt-core==1.6.18
dbt-duckdb==1.6.2
dbt-extractor==0.4.1
dbt-semantic-interfaces==0.2.3
docstring_parser==0.16
duckdb==1.3.1
exceptiongroup==1.3.0
fsspec==2025.5.1
gql==3.5.3
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
grpcio==1.73.1
grpcio-health-checking==1.62.3
h11==0.16.0
hologram==0.0.16
httptools==0.6.4
humanfriendly==10.0
idna==3.10
importlib-metadata==6.11.0
isodate==0.6.1
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
leather==0.4.0
Logbook==1.5.3
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mashumaro==3.8.1
mdurl==0.1.2
minimal-snowplow-tracker==0.0.2
more-itertools==8.14.0
msgpack==1.1.1
multidict==6.6.0
networkx==3.2.1
orjson==3.10.18
packaging==25.0
parsedatetime==2.6
pathspec==0.11.2
pendulum==3.1.0
polars==0.18.15
propcache==0.3.2
protobuf==4.25.8
pyarrow==20.0.0
pycparser==2.22
pydantic==1.10.22
Pygments==2.19.2
python-dateutil==2.8.2
python-dotenv==1.1.1
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2025.2
pytzdata==2020.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.25.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
starlette==0.47.1
structlog==25.4.0
tabulate==0.9.0
text-unidecode==1.3
tomli==2.2.1
toposort==1.10
tqdm==4.67.1
typer==0.16.0
typing_extensions==4.14.0
tzdata==2025.2
universal_pathlib==0.2.6
urllib3==1.26.20
uvicorn==0.34.3
uvloop==0.21.0
watchdog==6.0.0
watchfiles==1.1.0
websockets==15.0.1
yarl==1.20.1
zipp==3.23.0
